conformance: move kustomization logic into kustomize.go
